test: desacoplar testes de integração da infraestrutura de produção

+++ src/main/java/banco_api/dto/CriarContaCorrenteDTO.java
@@ -4,12 +4,15 @@ import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 @Setter
 public class CriarContaCorrenteDTO {
     @Positive(message = "O limite deve ser maior que zero")
-    private double limite;
+    private BigDecimal limite;
     @Positive(message = "O ID não pode ser negativo.")
     private Long pessoaId;
     @Positive(message = "O numero deve ser maior que zero.")
