+++ src/main/java/banco_api/dto/CriarContaPoupancaDTO.java
@@ -4,6 +4,9 @@ import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 @Setter
@@ -12,7 +15,7 @@ public class CriarContaPoupancaDTO {
     @Positive(message = "O número deve ser maior que zero.")
     private Long numero;
     @Positive(message = "A taxa de rendimento deve ser maior que zero.")
-    private double taxaRendimento;
+    private BigDecimal taxaRendimento;
     @Positive(message = "O ID não pode ser negativo.")
     private Long pessoaId;
