+++ src/main/java/banco_api/dto/DepositoDTO.java
@@ -4,10 +4,12 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 
 public class DepositoDTO {
     @Positive(message = "O valor deve ser maior que zero.")
-    private double valor;
+    private BigDecimal valor;
 }
